Remove the `enable-global-instrumentation` and `remove-global-instrumentation` commands
