+++ tracer/src/Datadog.FleetInstaller/Program.cs
@@ -38,12 +38,6 @@ rootCommand.AddExample("""
                        remove-iis-instrumentation"
                        """);
 rootCommand.AddExample("""
-                       enable-global-instrumentation --home-path "C:\datadog\versions\3.9.0"
-                       """);
-rootCommand.AddExample("""
-                       remove-global-instrumentation"
-                       """);
-rootCommand.AddExample("""
                        available-commands"
                        """);
 
@@ -51,8 +45,6 @@ rootCommand.AddCommand(new InstallVersionCommand());
 rootCommand.AddCommand(new UninstallVersionCommand());
 rootCommand.AddCommand(new EnableIisInstrumentationCommand());
 rootCommand.AddCommand(new RemoveIisInstrumentationCommand());
-rootCommand.AddCommand(new EnableGlobalInstrumentationCommand());
-rootCommand.AddCommand(new RemoveGlobalInstrumentationCommand());
 rootCommand.AddCommand(new AvailableCommandsCommand(rootCommand));
 
 return builder.Build().Invoke(args);
